refactor(cli): extract session lifecycle (new/fork/worktree/recreate) to session_cli.py (#495) (#532)
* refactor(cli): extract session lifecycle (new/fork/worktree/recreate) to session_cli.py (#495)

Phase 1 (Wave 1) of #495. Moves the session-lifecycle commands
(new, fork, worktree, recreate, session-defaults) plus their
domain-private worktree-registry helpers into agentwire/session_cli.py,
wired via register_session_parser through the _REGISTRARS loop. Shared
helpers imported from core; pure relocation, CLI surface unchanged.

Rebased onto the post-panes cli-extraction base.

Built by dotdev.dev

* test: point worktree integration test at session_cli (#495)
