+++ tests/unit/test_cli_commands.py
@@ -247,9 +247,9 @@ class TestCmdSendWaitReady:
 
 class TestCmdNewFirstMessage:
     def test_remote_rejected(self, capsys, monkeypatch):
-        from agentwire.__main__ import cmd_new
+        from agentwire.session_cli import cmd_new
 
-        monkeypatch.setattr("agentwire.__main__._check_tmux_installed", lambda: True)
+        monkeypatch.setattr("agentwire.session_cli._check_tmux_installed", lambda: True)
         args = argparse.Namespace(
             session="proj@gpu", path=None, force=False, json=True,
             roles=None, no_soul=True, first_message="an idea",
@@ -283,7 +283,8 @@ def _patch_role_pipeline(monkeypatch, projects_dir, project_config_roles):
     """
     from types import SimpleNamespace
 
-    import agentwire.__main__ as mod
+    import agentwire.session_cli as mod
+    from agentwire.core import AgentCommand
 
     cap = _RoleCapture()
 
@@ -299,7 +300,7 @@ def _patch_role_pipeline(monkeypatch, projects_dir, project_config_roles):
     })
     monkeypatch.setattr(mod, "load_project_config", lambda p: cfg)
     monkeypatch.setattr(mod, "detect_default_agent_type", lambda: "claude")
-    monkeypatch.setattr(mod, "build_agent_command", lambda *a, **k: mod.AgentCommand(command=""))
+    monkeypatch.setattr(mod, "build_agent_command", lambda *a, **k: AgentCommand(command=""))
 
     def fake_ensure_worktree(base, branch, wt, **kw):
         Path(wt).mkdir(parents=True, exist_ok=True)
@@ -340,7 +341,7 @@ class TestRecreateRoutesThroughResolveRoles:
     def test_worktree_recreate_reinjects_etiquette_even_without_saved_roles(
         self, monkeypatch, tmp_path
     ):
-        import agentwire.__main__ as mod
+        import agentwire.session_cli as mod
 
         projects = tmp_path / "projects"
         (projects / "proj").mkdir(parents=True)
@@ -359,7 +360,7 @@ class TestRecreateRoutesThroughResolveRoles:
     def test_worktree_recreate_stacks_saved_roles_under_etiquette(
         self, monkeypatch, tmp_path
     ):
-        import agentwire.__main__ as mod
+        import agentwire.session_cli as mod
 
         projects = tmp_path / "projects"
         (projects / "proj").mkdir(parents=True)
@@ -375,7 +376,7 @@ class TestRecreateRoutesThroughResolveRoles:
         assert "domain" in cap.role_names
 
     def test_plain_recreate_is_orchestrator_replaceable(self, monkeypatch, tmp_path):
-        import agentwire.__main__ as mod
+        import agentwire.session_cli as mod
 
         projects = tmp_path / "projects"
         (projects / "proj").mkdir(parents=True)
@@ -389,7 +390,7 @@ class TestRecreateRoutesThroughResolveRoles:
         assert "custom" in cap.role_names
 
     def test_plain_recreate_zero_config_is_orchestrator(self, monkeypatch, tmp_path):
-        import agentwire.__main__ as mod
+        import agentwire.session_cli as mod
 
         projects = tmp_path / "projects"
         (projects / "proj").mkdir(parents=True)
@@ -403,7 +404,7 @@ class TestRecreateRoutesThroughResolveRoles:
 
 class TestForkRoutesThroughResolveRoles:
     def test_worktree_fork_injects_worktree_session_etiquette(self, monkeypatch, tmp_path):
-        import agentwire.__main__ as mod
+        import agentwire.session_cli as mod
 
         projects = tmp_path / "projects"
         (projects / "proj").mkdir(parents=True)  # source_path (no source branch)
@@ -419,7 +420,7 @@ class TestForkRoutesThroughResolveRoles:
         assert cap.role_names[0] == "worktree-session"
 
     def test_worktree_fork_stacks_source_roles_under_etiquette(self, monkeypatch, tmp_path):
-        import agentwire.__main__ as mod
+        import agentwire.session_cli as mod
 
         projects = tmp_path / "projects"
         (projects / "proj").mkdir(parents=True)
@@ -435,7 +436,7 @@ class TestForkRoutesThroughResolveRoles:
         assert "domain" in cap.role_names
 
     def test_non_worktree_fork_is_orchestrator_replaceable(self, monkeypatch, tmp_path):
-        import agentwire.__main__ as mod
+        import agentwire.session_cli as mod
 
         projects = tmp_path / "projects"
         projects.mkdir(parents=True)
